+++ xmake/modules/private/action/require/impl/actions/download_resources.lua
@@ -114,7 +114,9 @@ function _download(package, resource_name, resource_url, resource_hash)
             -- we can use local resource from the search directories directly if network is too slow
             os.cp(localfile, resource_file)
         elseif resource_url:find(string.ipattern("https-://")) or resource_url:find(string.ipattern("ftps-://")) then
-            http.download(resource_url, resource_file, {insecure = global.get("insecure-ssl")})
+            http.download(resource_url, resource_file, {
+                insecure = global.get("insecure-ssl"),
+                headers = package:policy("package.download.http_headers")})
         else
             raise("invalid resource url(%s)", resource_url)
         end